Add Android platform; Import testify; Fix tests
I've added Android to the list of available platorms, as it was breaking tests.
I think this is the recent PR to tldr-pages that added Android:
https://github.com/tldr-pages/tldr/pull/5743

Also, I've imported the testify lib, as I'm using it extensively, and it just
makes my life easier. :)
https://github.com/stretchr/testify

Finally, I refactored the TestAvailablePlatforms to table driven tests.